# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FinanceDOM API — Subventions non remboursables Outre-Mer
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dossiers/{id}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /dossiers/{dossierId}/documents/{docId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - name: dossierId
          in: path
          required: true
          schema:
            type: integer
        - name: docId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Messages ──────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──
...
    Document:
      type: object
      properties:
        id:
          type: integer
        dossierId:
          type: integer
        type:
          type: string
        nom:
          type: string
        filename:
          type: string
        statut:
          type: string
          enum: [en_attente, valide, rejete, manquant]
        obligatoire:
          type: boolean
        uploadedAt:
          type: string
      required: [id, dossierId, type, nom, statut, obligatoire]
    UploadDocumentBody:
      type: object
      properties:
        type:
          type: string
        nom:
          type: string
        filename:
          type: string
      required: [type, nom, filename]

    # ── Messages ──
...